refactor: extract shared Motion console bench base
Move FpgaRegisterIO, the console FPGA revision reader, and the ten
console/identity/config/trigger methods duplicated between
MotionLaserCalibrationBench and MotionSafetyCalibrationBench into a new
meter-agnostic motion_bench module. laser_hardware re-exports the moved
names for existing importers; safety_hardware stays Ophir-free.
